+++ Quiper/MainWindowController+Actions.swift
@@ -209,7 +209,7 @@ extension MainWindowController {
     @objc func refreshStopTapped(_ sender: NSButton) {
         guard let webView = currentWebView() else { return }
         if webView.isLoading {
-            webView.stopLoading()
+            webViewManager.stopLoading(webView)
         } else {
             webView.reload()
         }